Allow user to use app without login

+++ src/context/TodoContext.js
@@ -1,4 +1,10 @@
-import { createContext, useContext, useEffect, useReducer } from 'react';
+import {
+  createContext,
+  useContext,
+  useEffect,
+  useReducer,
+  useRef,
+} from 'react';
 import { getAuth, onAuthStateChanged, updateProfile } from 'firebase/auth';
 import {
   collection,
@@ -42,6 +48,7 @@ const TodoProvider = ({ children }) => {
   const [state, dispatch] = useReducer(todoReducer, initialState);
 
   const auth = getAuth();
+  const localStorageRef = useRef(false);
 
   // Set current user
   useEffect(() => {
@@ -56,6 +63,9 @@ const TodoProvider = ({ children }) => {
   // Get user's todos
   useEffect(() => {
     const getTodos = async () => {
+      let todos;
+
+      // If signed in, get from database
       if (state.currentUser) {
         const q = query(
           collection(db, 'todos'),
@@ -63,29 +73,44 @@ const TodoProvider = ({ children }) => {
         );
 
         const querySnap = await getDocs(q);
-        const todos = [];
+        todos = [];
         querySnap.forEach((doc) => {
           todos.push({
             ...doc.data(),
           });
         });
-
-        dispatch({
-          type: GET_TODOS,
-          payload: todos,
-        });
+      } else {
+        // If not, get from local storage
+        todos = JSON.parse(localStorage.getItem('todos')) || [];
       }
+
+      dispatch({
+        type: GET_TODOS,
+        payload: todos,
+      });
     };
 
     getTodos();
   }, [state.currentUser]);
+
+  // Update todos in local storage if not signed in
+  useEffect(() => {
+    if (!state.currentUser) {
+      // Ignore state's todos in first render
+      if (!localStorageRef.current) {
+        localStorageRef.current = true;
+      } else {
+        localStorage.setItem('todos', JSON.stringify(state.todos));
+      }
+    }
+  }, [state.todos, state.currentUser]);
 
   const addTodo = async (task) => {
     const newTodo = {
       id: uuidv4(),
       task,
       completed: false,
-      userRef: state.currentUser.uid,
+      userRef: state.currentUser?.uid,
     };
 
     dispatch({
@@ -93,7 +118,9 @@ const TodoProvider = ({ children }) => {
       payload: newTodo,
     });
 
-    await setDoc(doc(db, 'todos', newTodo.id), newTodo);
+    if (state.currentUser) {
+      await setDoc(doc(db, 'todos', newTodo.id), newTodo);
+    }
   };
 
   const checkTodo = async (id) => {
@@ -102,12 +129,13 @@ const TodoProvider = ({ children }) => {
       payload: id,
     });
 
-    await updateDoc(doc(db, 'todos', id), {
-      completed: true,
-    });
-
-    // Update completed count
     if (state.currentUser) {
+      // Mark task as completed
+      await updateDoc(doc(db, 'todos', id), {
+        completed: true,
+      });
+
+      // Update total completed tasks count
       const currentUser = await getDoc(doc(db, 'users', state.currentUser.uid));
       const { completed } = currentUser.data();
 
@@ -123,12 +151,13 @@ const TodoProvider = ({ children }) => {
       payload: id,
     });
 
-    await updateDoc(doc(db, 'todos', id), {
-      completed: false,
-    });
-
-    // Update completed count
     if (state.currentUser) {
+      // Mark task as incompleted
+      await updateDoc(doc(db, 'todos', id), {
+        completed: false,
+      });
+
+      // Update total completed tasks count
       const currentUser = await getDoc(doc(db, 'users', state.currentUser.uid));
       const { completed } = currentUser.data();
 
@@ -144,7 +173,9 @@ const TodoProvider = ({ children }) => {
       payload: id,
     });
 
-    await deleteDoc(doc(db, 'todos', id));
+    if (state.currentUser) {
+      await deleteDoc(doc(db, 'todos', id));
+    }
   };
 
   const showAllTodos = () => {
@@ -160,13 +191,15 @@ const TodoProvider = ({ children }) => {
   };
 
   const clearCompletedTodos = async () => {
-    const completedTodos = state.todos.filter((todo) => todo.completed);
-
-    completedTodos.forEach(
-      async (todo) => await deleteDoc(doc(db, 'todos', todo.id))
-    );
-
     dispatch({ type: CLEAR_COMPLETED });
+
+    if (state.currentUser) {
+      const completedTodos = state.todos.filter((todo) => todo.completed);
+
+      completedTodos.forEach(
+        async (todo) => await deleteDoc(doc(db, 'todos', todo.id))
+      );
+    }
   };
 
   const updateUserProfile = async (prop, val) => {
